Align TypeScript schemas with database schema
- Add compare-schema.ts script to detect schema drift
- Fix type mismatches: users.id, user_roles.id/userId (BIGINT→INTEGER), tracks.campaignId (INTEGER→BIGINT)
- Add missing columns to tracks (artLink, hookTimings, artworkLink)
- Add missing columns to playlists (created_by)
- Add missing columns to licenses (projectId, type)
- Update sound_projects schema (replace campaignIds with campaignId and add video/editing columns)
- Add missing columns to project_tracks (fromTime, tillTime, volume, finalVideoPath)
- Fix brands.restrictedOwners type (TEXT[]→STRING[])
- Fix faqs.createdAt nullability
- Update getUserUsedCampaignIds to use new campaignId field

+++ services/persistence-service/user/schemas/user-role.schema.ts
@@ -29,12 +29,12 @@ export interface UserRoleDetails {
 export class UserRoleModel extends Model<UserRoleModel, UserRoleDetails> {
   @PrimaryKey
   @AutoIncrement
-  @Column(DataType.BIGINT)
+  @Column(DataType.INTEGER)
   id!: number;
 
   @ForeignKey(() => UserModel)
   @Column({
-    type: DataType.BIGINT,
+    type: DataType.INTEGER,
     allowNull: false,
   })
   userId!: number;
